Recode object keys in Legacy Core bucket
This commit edits the core_file_constructor() and R scripts in
data_generation.R to read in the renamed legacy core objects on S3.
The objects were renamed to improve readability by removing the
redundant "SCOPE" string from file names. Thus, nccsdata's get_data()
had to be debugged to read from the correct S3 urls pointing to the
renamed object keys. Additionally, the internal named list used to store
object keys had to be reconstructed as well to account for the renamed
keys.