Add geo_preview() function and associated dependencies
This commit adds the geo_preview() function and its associated
dependencies to the code base. The function helps users parse the fips
codes needed to filter NCCS data. Filtering via FIPS also ensures that
geographic subsets are based on census codes and not other geographic
classifications. This leads to greater external validity for pulled data
and standardizes the units of analysis in downstream research. It also
supports the reproducibility of prior research.