Build vignette for geo_preview() function
This commit renders the html vignette for the geo_preview() function,
providing documentation for package users who want to retrieve
geographic metadata using geo_preview(). It also edits geo_preview() to
return an invisible return of the previewed dataframe while printing the
output with knitr for use in downstream tasks and readability
respectively.