Fix bug in get_bmf()
This commit fixes a bug in get_bmf() raised in Issue # 15.

The bug happens when appending the bmf to core files. bmf.EIN and
core.EIN are different data types hence merging cannot happen. To fix
this, EINs were standardized as characters with leading 0s imputed until
each EIN is standardized at 9 digits. This data type allows each EIN to
act as a unique identifier.

Credit to chickymonkeys for spotting this bug